pyproject.toml
[project]
name = "llmdata"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "faiss-cpu>=1.12.0",
    "hf-xet>=1.1.10",
    "langchain-community>=0.3.29",
    "langchain-google-genai>=2.1.12",
    "lxml>=6.0.2",
    "pandas>=2.3.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "ultimate-sitemap-parser>=1.6.0",
    "unstructured>=0.18.15",
]
